Add updating of subjectSets before running the export job
TMT-265
When autosave-subject-sets option was set to true, exported data was not
up to date with a selection from step1.

The export behaviour for autosave-subject-sets mode is changed to first
check if step 1 subject set is up to date with criteria. If it is not, a
new subject set is created and then based on it the export is created.